[spring-boot-kotlin] feature: mysql, postgresql 통합 테스트 환경 구성

+++ spring-boot-kts/test-support/testcontainers-support/src/main/kotlin/dev/pkch/support/DockerComposeContainerConfiguration.kt
@@ -60,7 +60,9 @@ abstract class DockerComposeContainerConfiguration(
 
     @PreDestroy
     fun stop() {
-        container.stop()
-        dockerComposeFile.delete()
+        if (container != null) {
+            container.stop()
+            dockerComposeFile.delete()
+        }
     }
 }
